requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
openai>=1.0.0
pydantic>=2.0.0
pymysql>=1.1.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
# uvicorn>=0.23.0
# mcp-server-api>=0.1.0
typing-extensions>=4.0.0